Test/controllers

+++ src/main/java/com/niallantony/deulaubaba/services/StudentService.java
@@ -48,20 +48,20 @@ public class StudentService {
 
     public StudentIdAvatar getStudentPreviewById(String id)  {
         Student student = studentRepository.findById(id.toLowerCase())
-                .orElseThrow(() -> new ResourceNotFoundException("Student not found" + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Student not found " + id));
         return getStudentIdAvatar(student);
     }
 
     public StudentDTO getStudentById(String id)  {
         Student student = studentRepository.findById(id.toLowerCase())
-                .orElseThrow(() -> new ResourceNotFoundException("Student not found" + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Student not found " + id));
         return studentMapper.toDTO(student);
     }
 
     public List<UserAvatar> getStudentTeam(String id)  {
 
-       Student student = studentRepository.findById(id)
-               .orElseThrow(() -> new ResourceNotFoundException("Student not found" + id));
+       Student student = studentRepository.findById(id.toLowerCase())
+               .orElseThrow(() -> new ResourceNotFoundException("Student not found " + id));
        Set<User> users = student.getUsers();
         return users.stream()
                .map(user -> new UserAvatar(
@@ -74,35 +74,20 @@ public class StudentService {
     public List<StudentIdAvatar> getStudents(String id)  {
         List<StudentIdAvatar> students = studentRepository.findAllOfUserId(id);
         if (students.isEmpty()) {
-            throw new ResourceNotFoundException("Students not found" + id);
+            throw new ResourceNotFoundException("Students not found");
         }
         return students;
     }
 
-    private Student extractStudent(StudentRequest studentRequest, String userId) {
+    private Student extractStudent(StudentRequest studentRequest, User user) {
 
         char[] alphabet = {'1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'};
         Random random = new Random();
-        User user = userRepository.findByUserId(userId).orElseThrow(() -> new UserNotAuthorizedException("Unauthorized Access"));
-
         Student student = studentMapper.toStudent(studentRequest);
         student.setStudentId(NanoIdUtils.randomNanoId(random, alphabet, 6));
         student.getUsers().add(user);
 
         return student;
-    }
-
-    @Transactional
-    public StudentDTO createStudent(String request, String userId) {
-        StudentRequest studentRequest = jsonUtils.parse(
-                request,
-                StudentRequest.class,
-                () -> new InvalidStudentDataException("Invalid request")
-        );
-        Student student = extractStudent(studentRequest, userId);
-        studentRepository.save(student);
-        linkStudentToUser(student, userId);
-        return studentMapper.toDTO(student);
     }
 
     @Transactional
@@ -112,32 +97,21 @@ public class StudentService {
                 StudentRequest.class,
                 () -> new InvalidStudentDataException("Invalid request")
         );
-        Student student = extractStudent(studentRequest, userId);
-        try {
-            String filename = fileStorageService.storeImage(image);
-            student.setImagesrc(filename);
-        } catch (FileStorageException e) {
-            log.warn("Image not saved to storage", e);
+        User user = userRepository.findByUserId(userId).orElseThrow(() -> new UserNotAuthorizedException("Unauthorized access"));
+        validateStudentRequest(studentRequest);
+        Student student = extractStudent(studentRequest, user);
+        if (image != null && !image.isEmpty()) {
+            try {
+                String filename = fileStorageService.storeImage(image);
+                student.setImagesrc(filename);
+            } catch (FileStorageException e) {
+                log.warn("Image not saved to storage", e);
+            }
         }
         studentRepository.save(student);
         linkStudentToUser(student, userId);
 
         return studentMapper.toDTO(student);
-    }
-
-    @Transactional
-    public StudentDTO updateStudentDetails(String studentId, String request, String userId)  {
-        StudentRequest studentRequest = jsonUtils.parse(
-                request,
-                StudentRequest.class,
-                () -> new InvalidStudentDataException("Invalid request")
-        );
-        Student student = getAuthorisedStudent(studentId, userId);
-
-       applyDetailUpdates(student, studentRequest);
-
-       studentRepository.save(student);
-       return studentMapper.toDTO(student);
     }
 
     // TODO: Image storage failure should be told to user
@@ -150,23 +124,27 @@ public class StudentService {
                 () -> new InvalidStudentDataException("Invalid request")
         );
         Student student = getAuthorisedStudent(studentId, userId);
-        String oldImg = student.getImagesrc();
-
+        String oldImg = null;
         boolean newImageStored = false;
-        try {
-            String filename = fileStorageService.storeImage(image);
-            student.setImagesrc(filename);
-            newImageStored = true;
-        } catch (FileStorageException e) {
-            log.warn("Image not saved to storage", e);
+        if (image != null && !image.isEmpty()) {
+            oldImg = student.getImagesrc();
+            try {
+                String filename = fileStorageService.storeImage(image);
+                student.setImagesrc(filename);
+                newImageStored = true;
+            } catch (FileStorageException e) {
+                log.warn("Image not saved to storage", e);
+            }
         }
         applyDetailUpdates(student, studentRequest);
-
         studentRepository.save(student);
-        if (newImageStored) {
-            fileStorageService.deleteImage(oldImg);
+        if (newImageStored && oldImg != null) {
+            try {
+                fileStorageService.deleteImage(oldImg);
+            } catch (FileStorageException e) {
+                log.warn("Couldn't delete old image from storage", e);
+            }
         }
-
         return studentMapper.toDTO(student);
     }
 
@@ -196,18 +174,20 @@ public class StudentService {
         return studentMapper.toDTO(student);
     }
 
+    @Transactional
     public Student getAuthorisedStudent(String studentId, String userId) {
         Student student = studentRepository.findById(studentId)
                 .orElseThrow(() -> new ResourceNotFoundException("Student not found " + studentId));
         User user = userRepository.findByUserId(userId)
                 .orElseThrow(() -> new UserNotAuthorizedException("User not found " + userId));
         if (!student.getUsers().contains(user)) {
-            throw new UserNotAuthorizedException("Unauthorized Access");
+            throw new UserNotAuthorizedException("Unauthorized access");
         }
         return student;
     }
     
     private void applyDetailUpdates(Student student, StudentRequest studentRequest) {
+        validateStudentRequest(studentRequest);
         student.setName(studentRequest.getName());
         student.setSchool(studentRequest.getSchool());
         student.setAge(studentRequest.getAge());
@@ -225,7 +205,7 @@ public class StudentService {
     }
 
     public boolean studentBelongsToUser(String studentId, String userId) {
-        Student student = studentRepository.findById(studentId)
+        Student student = studentRepository.findById(studentId.toLowerCase())
                 .orElseThrow(() -> new ResourceNotFoundException("Student not found " + studentId));
         User user = userRepository.findByUserId(userId)
                 .orElseThrow(() -> new UserNotAuthorizedException("User not found " + userId));
@@ -238,6 +218,20 @@ public class StudentService {
         user.getStudents().add(student);
         student.getUsers().add(user);
         userRepository.save(user);
+    }
+
+    private void validateStudentRequest(StudentRequest studentRequest) {
+        if (
+                studentRequest.getName() == null
+                || studentRequest.getSchool() == null
+                || studentRequest.getAge() == null
+                || studentRequest.getGrade() == null
+                || studentRequest.getDisability() == null
+                || studentRequest.getSetting() == null
+        ) {
+            throw new InvalidStudentDataException("Invalid student data");
+        }
+
     }
 
 }
